# openapi.json
{
  "components": {
    "responses": {
...
      "InvalidToken": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "The token is invalid or expired, or its user is inactive (`\"Invalid token\"`). The UI logs the user out on this response."
      },
...
      "RateLimited": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "Too many requests from this IP (`\"Too many requests. Please try again later.\"`). `RateLimit-*` headers describe the window."
      },
      "ServerError": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "An unexpected error (`\"Something went wrong. Please try again.\"`). Details are logged on the server, never returned."
      },
      "TokenNotFound": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "No `Authorization` header was sent (`\"Token not found\"`)."
      },
...
    },
    "schemas": {
...
      "BookSearchResponse": {
        "properties": {
          "data": {
            "properties": {
              "mediaList": {
                "items": {
                  "$ref": "#/components/schemas/BookSearchResult"
                },
                "type": "array"
              },
              "totalPages": {
                "type": "integer"
              }
            },
            "required": [
              "mediaList",
              "totalPages"
            ],
            "type": "object"
          },
          "mediaType": {
            "enum": [
              "book"
            ],
            "type": "string"
          },
          "status": {
            "enum": [
              "success"
            ],
            "type": "string"
          }
        },
        "required": [
          "status",
          "data",
          "mediaType"
        ],
        "type": "object"
      },
      "BookSearchResult": {
        "properties": {
          "author": {
            "description": "Author names joined with commas.",
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "mediaId": {
            "description": "Google Books volume id.",
            "type": "string"
          },
          "mediaType": {
            "enum": [
              "book"
            ],
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "poster": {
            "type": "string"
          }
        },
        "required": [
          "mediaId",
          "mediaType"
        ],
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "errors": {
            "properties": {
              "msg": {
                "example": "User not found",
                "type": "string"
              }
            },
            "required": [
              "msg"
            ],
            "type": "object"
          }
        },
        "required": [
          "errors"
        ],
        "type": "object"
      },
...
      "MovieTvSearchResponse": {
        "properties": {
          "data": {
            "properties": {
              "mediaList": {
                "items": {
                  "$ref": "#/components/schemas/MovieTvSearchResult"
                },
                "type": "array"
              },
              "totalPages": {
                "type": "integer"
              }
            },
            "required": [
              "mediaList",
              "totalPages"
            ],
            "type": "object"
          },
          "mediaType": {
            "enum": [
              "movie",
              "tv"
            ],
            "type": "string"
          },
          "status": {
            "enum": [
              "success"
            ],
            "type": "string"
          }
        },
        "required": [
          "status",
          "data",
          "mediaType"
        ],
        "type": "object"
      },
      "MovieTvSearchResult": {
        "properties": {
          "description": {
            "type": "string"
          },
          "mediaId": {
            "description": "TMDB id.",
            "type": "integer"
          },
          "mediaType": {
            "enum": [
              "movie",
              "tv"
            ],
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "poster": {
            "description": "Poster URL, or an empty string when TMDB has none.",
            "type": "string"
          }
        },
        "required": [
          "mediaId",
          "poster",
          "mediaType"
        ],
        "type": "object"
      },
      "MusicSearchResponse": {
        "properties": {
          "data": {
            "properties": {
              "mediaList": {
                "items": {
                  "$ref": "#/components/schemas/MusicSearchResult"
                },
                "type": "array"
              },
              "totalPages": {
                "type": "integer"
              }
            },
            "required": [
              "mediaList",
              "totalPages"
            ],
            "type": "object"
          },
          "mediaType": {
            "enum": [
              "music"
            ],
            "type": "string"
          },
          "status": {
            "enum": [
              "success"
            ],
            "type": "string"
          }
        },
        "required": [
          "status",
          "data",
          "mediaType"
        ],
        "type": "object"
      },
      "MusicSearchResult": {
        "properties": {
          "albumName": {
            "type": "string"
          },
          "albumType": {
            "type": "string"
          },
          "artists": {
            "description": "Artist names joined with commas.",
            "type": "string"
          },
          "mediaId": {
            "description": "Spotify track id.",
            "type": "string"
          },
          "mediaType": {
            "enum": [
              "music"
            ],
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "poster": {
            "description": "Set only when Spotify has a 640px album image.",
            "type": "string"
          }
        },
        "required": [
          "mediaId",
          "artists",
          "mediaType"
        ],
        "type": "object"
      },
      "ObjectId": {
        "example": "63019b905ccf53564ffb0c84",
        "pattern": "^[a-f0-9]{24}$",
        "type": "string"
      },
      "PublicUser": {
        "properties": {
          "_id": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "firstName": {
            "type": "string"
          },
          "followers": {
            "description": "userNames of followers.",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "following": {
            "description": "userNames this user follows.",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "lastName": {
            "type": "string"
          },
          "picture": {
            "description": "Profile picture URL.",
            "type": "string"
          },
          "userName": {
            "description": "Missing until a social sign-up completes their profile.",
            "type": "string"
          }
        },
        "required": [
          "_id"
        ],
        "type": "object"
      },
...
      "SearchRequest": {
        "properties": {
          "keyWord": {
            "type": "string"
          },
          "page": {
            "description": "1-based page number. Defaults to the first page.",
            "minimum": 1,
            "type": "integer"
          }
        },
        "required": [
          "keyWord"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "tokenAuth": {
        "description": "The JWT returned by signup/login, sent as the raw header value — no `Bearer ` prefix.",
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "API for Rate It, a social app for rating movies, TV, music, and books. Success responses are `{ status: \"success\", data: { ... } }`; errors are `{ errors: { msg } }`.",
    "title": "Rate It API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/search/all": {
      "post": {
        "description": "First page of each catalog. A catalog that fails comes back as an empty list instead of failing the whole request.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "keyWord": {
                    "type": "string"
                  }
                },
                "required": [
                  "keyWord"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "properties": {
                        "fullSearchList": {
                          "properties": {
                            "book": {
                              "items": {
                                "$ref": "#/components/schemas/BookSearchResult"
                              },
                              "type": "array"
                            },
                            "movie": {
                              "items": {
                                "$ref": "#/components/schemas/MovieTvSearchResult"
                              },
                              "type": "array"
                            },
                            "music": {
                              "items": {
                                "$ref": "#/components/schemas/MusicSearchResult"
                              },
                              "type": "array"
                            },
                            "tv": {
                              "items": {
                                "$ref": "#/components/schemas/MovieTvSearchResult"
                              },
                              "type": "array"
                            }
                          },
                          "required": [
                            "movie",
                            "tv",
                            "music",
                            "book"
                          ],
                          "type": "object"
                        }
                      },
                      "required": [
                        "fullSearchList"
                      ],
                      "type": "object"
                    },
                    "status": {
                      "enum": [
                        "success"
                      ],
                      "type": "string"
                    }
                  },
                  "required": [
                    "status",
                    "data"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Results grouped by catalog."
          },
          "403": {
            "$ref": "#/components/responses/InvalidToken"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "security": [
          {},
          {
            "tokenAuth": []
          }
        ],
        "summary": "Search movies, TV, music, and books at once",
        "tags": [
          "Search"
        ]
      }
    },
    "/api/search/book": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookSearchResponse"
                }
              }
            },
            "description": "One page of results (20 per page)."
          },
          "403": {
            "$ref": "#/components/responses/InvalidToken"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`\"Unable to search books\"`."
          }
        },
        "security": [
          {},
          {
            "tokenAuth": []
          }
        ],
        "summary": "Search books on Google Books",
        "tags": [
          "Search"
        ]
      }
    },
    "/api/search/movie": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovieTvSearchResponse"
                }
              }
            },
            "description": "One page of results."
          },
          "403": {
            "$ref": "#/components/responses/InvalidToken"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`\"Unable to search movies\"`."
          }
        },
        "security": [
          {},
          {
            "tokenAuth": []
          }
        ],
        "summary": "Search movies on TMDB",
        "tags": [
          "Search"
        ]
      }
    },
    "/api/search/music": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MusicSearchResponse"
                }
              }
            },
            "description": "One page of results (20 per page)."
          },
          "403": {
            "$ref": "#/components/responses/InvalidToken"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`\"Unable to search music\"`."
          }
        },
        "security": [
          {},
          {
            "tokenAuth": []
          }
        ],
        "summary": "Search songs on Spotify",
        "tags": [
          "Search"
        ]
      }
    },
    "/api/search/tv": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovieTvSearchResponse"
                }
              }
            },
            "description": "One page of results."
          },
          "403": {
            "$ref": "#/components/responses/InvalidToken"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`\"Unable to search TV shows\"`."
          }
        },
        "security": [
          {},
          {
            "tokenAuth": []
          }
        ],
        "summary": "Search TV shows on TMDB",
        "tags": [
          "Search"
        ]
      }
    },
    "/api/search/user": {
      "post": {
        "description": "Case-insensitive. Returns every match with no paging; `page` is ignored. Results come back under `mediaList`, like the media searches.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "keyWord": {
                    "type": "string"
                  }
                },
                "required": [
                  "keyWord"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "properties": {
                        "mediaList": {
                          "items": {
                            "$ref": "#/components/schemas/PublicUser"
                          },
                          "type": "array"
                        }
                      },
                      "required": [
                        "mediaList"
                      ],
                      "type": "object"
                    },
                    "mediaType": {
                      "enum": [
                        "user"
                      ],
                      "type": "string"
                    },
                    "status": {
                      "enum": [
                        "success"
                      ],
                      "type": "string"
                    }
                  },
                  "required": [
                    "status",
                    "data",
                    "mediaType"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Matching users."
          },
          "401": {
            "$ref": "#/components/responses/TokenNotFound"
          },
          "403": {
            "$ref": "#/components/responses/InvalidToken"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "security": [
          {
            "tokenAuth": []
          }
        ],
        "summary": "Search users by userName or first name",
        "tags": [
          "Search"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
}